chore: clean home component
